Reformat of .py files to move from old application.py to new files: config.py __init__.py blog.py views.py dbmodels.py
deletion of old application.py file
deletion of old controllers.py
deletion of old database_tables.py
