+++ app/templates/base.html
@@ -29,10 +29,10 @@
           <a class="nav-link" href="{{url_for('index')}}">Home</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{url_for('catalog')}}">Catalog</a>
+          <a class="nav-link" href="{{url_for('catalog')}}">Plant Catalog</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{url_for('calendar')}}">Calendar</a>
+          <a class="nav-link" href="{{url_for('calendar')}}">Planting Calendar</a>
         </li>
         <li class="nav-item">
           <a class="nav-link" href="{{url_for('search')}}">Plant Search</a>
