+++ app/templates/plantinfo.html
