+++ app/templates/privacy.html
