+++ blog.py
@@ -1,0 +1,3 @@
+from app import app
+
+app.run(debug=True)
