--- database_tables.py
@@ -1,14 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from application import db
-
-
-def __repr__(self):
-    return '<PlantName {}>'.format(self.plant_name)
-
-
-
-
-
-
-
