--- templates/calendar.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-
-
-{% block title %} Links {% endblock %}
-
-
-{% block body %}
-
-<h1> Calendar </h1>
-
-<p> Testing </p>
-
-{% endblock %}
