--- templates/search.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-
-
-{% block title %} Search {% endblock %}
-
-
-{% block body %}
-
-<h1> Search the Catalog </h1>
-
-<p> Testing </p>
-
-{% endblock %}
